Change placement handling
Variables are added to the robot and used to compute the robot's forward
kinematics. Links and joints now have one placement, `Placement`, for
the placement relative to the robot's base link. Links have another
placement, `MountedPlacement` (for now), which corresponds to `origin`
in URDF. This is the user's responsibility to correctly set
`MountedPlacement` for the real, visual, and collision parts.