Improve handling of mobile robots in workcell
Do not add the root link if it is not set in the workcell and there is
only one XacroObject without parent. This allows to use mobile robots
not fixed to anything.
Add a root link if there is more than one XacroObject without parent and
warn the user. This allows to generate a functional xacro.