Handle block parameters in xacro macros
There is still an issue because block parameters are empty when loading a
xacro file.
The workbench replaces the block parameters with a comment but this is
not a valid URDF tag.
The issue is solved when the user gives a valid value to the block
parameter but it would be better to handle also the empty case.