Fix `attached_collision_object.svg`
Inkscape doesn't use the `fill` attribute when the `style` attribute is
present but FreeCAD (QT) does and gives it a higher priority than
`style`.
Icons thus may look differently in Inkscape and FreeCAD.
This should be fixed for all icons...

By the way, copying the style in Inkscape doesn't remove the `fill`
attribute but probably only copies the `style` attribute.

Signed-off-by: Gaël Écorchard <gael@km-robotics.cz>